Split Op and Block
"log.rs" is a confusing name, they should really be in their own
separate modules.
