+++ src/proto/block.proto
@@ -1,18 +1,6 @@
 package ithos;
 
-import "objectclass.proto";
-
-// Ops make modifications to the state of the database
-message Op {
-  enum Type {
-    ADD = 1;
-  }
-
-  required Type        optype = 1;
-  required string      path = 2;
-  optional ObjectClass objectclass = 3;
-  optional bytes       data = 4;
-}
+import "op.proto";
 
 // Blocks are the fundamental replication unit of Ithos, and consist of
 // multiple operations plus the cryptographic authority (in the form of
